requirements.txt
flask-mysqldb==0.2.0
python-dotenv==1.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0